requirements.txt
streamlit
firebase-admin
google-generativeai
plotly
python-dotenv
streamlit-option-menu
